% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derappp-package.R
\docType{package}
\name{derappp-package}
\alias{derappp-package}
\title{derappp: Data for Environmental Risk Assessment of Plant Protection Products}
\description{
Makes some public data for environmental risk assessment of plant protection products accessible in R. Includes definitions of substances in terms of well-defined chemical entities, if applicable. Also includes source references including page numbers, most often referring to secondary sources, as the original reports are usually unpublished. Uses standardised scientific names for biological species and includes physicochemical as well as ecotoxicological data. Data sources are mainly conclusions from the European pesticide risk assessment peer review process published by the European Food Safety Authority (EFSA) and Renewal Assessment Reports (RARs), also published by EFSA. Note that the use of the data included in this package for registration purposes may be restricted by regulatory data protection, also known as data exclusivity rules, as detailed for example in article 59 of Regulation (EC) No 1107/2009.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://agroscope-ch.github.io/derappp/}
  \item Report bugs at \url{https://github.com/agroscope-ch/derappp/issues}
}

}
\author{
\strong{Maintainer}: Johannes Ranke \email{johannes.ranke@agroscope.admin.ch} (\href{https://orcid.org/0000-0003-4371-6538}{ORCID}) (acronym: rajo)

Authors:
\itemize{
  \item Johannes Ranke \email{johannes.ranke@agroscope.admin.ch} (\href{https://orcid.org/0000-0003-4371-6538}{ORCID}) (acronym: rajo)
  \item Elisabeth Lutz \email{elisabeth.lutz@agroscope.admin.ch} (acronym: luel, data generation scripts, aquatic toxicity data)
  \item Marcel Mathis \email{marcel.mathis@agroscope.admin.ch} (acronym: mamr, data generation scrips, aquatic toxicity data)
}

Other contributors:
\itemize{
  \item Laurent Boualit (acronym: bolr, aquatic toxicity data) [contributor]
  \item Jhoniel Perdigon Ferreira (acronym: pejh, data generation scripts, soil toxicity data, review of aquatic toxicity data) [contributor]
  \item Kasteel Romualdus \email{romualdus.kasteel@agroscope.admin.ch} (Reviewed the initial version before publication) [reviewer]
  \item Daniel Baumgartner (SMILES codes for pyrethroid stereoisomers) [contributor]
  \item Fabian Balk (\href{https://orcid.org/0000-0001-7764-7214}{ORCID}) (acronym: bafa, review of soil toxicity data) [reviewer]
  \item European Food Safety Authority (EFSA conclusions, bibliographic information in OpenFoodTox, Renewal Assessment Reports) [data contributor]
  \item British Crop Protection Council (Chemical identity information from the Compendium of Pesticide Common Names) [data contributor]
  \item National Center for Biotechnology Information (Chemical identity information retrieved from the PubChem website) [data contributor]
  \item CAS division of the American Chemical Society (Chemical identity information retreived from the CAS Common Chemistry website) [data contributor]
  \item enviPath UG & Co KG (Chemical identity information retreived from the envipath website) [data contributor]
  \item Agroscope [copyright holder]
}

}
\keyword{internal}
