% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derappp.R
\docType{data}
\name{derappp}
\alias{derappp}
\title{Data for environmental risk assessment of plant protection products}
\format{
A number of tables collected in a \link[dm:dm]{dm::dm} object
}
\usage{
derappp
}
\description{
Data for environmental risk assessment of plant protection products
}
\examples{
library(derappp)
library(dplyr, warn.conflicts = FALSE)
dm::dm_draw(derappp)

# The list of chemical entities ("chents")
derappp$chents

# Some vapor pressures and water solubilities
# We need to convert to °C if we do not want Kelvin
library(units)
derappp$p0[1:2, ] |>
  mutate(T = set_units(T, "°C"))
derappp$p0[1, ] |>
  left_join(derappp$sources, by = "sk") |>
  select(substance, p0, T, reference)

derappp$cwsat[1:3, ] |>
  left_join(derappp$sources, by = "sk") |>
  select(substance, cwsat, T, pH, reference) |>
  mutate(cwsat = set_units(cwsat, "g/L")) |>
  mutate(T = set_units(T, "°C"))

# Join names used in the Swiss register
derappp$chents |>
  left_join(derappp$substance_keys, by = c(chent = "substance")) |>
  filter(db == "srppp") |>
  left_join(srppphist::srppp_active_substances, by = c(key = "pk")) |>
  select(chent, iso, smiles, substance_de)

# Show some soil sorption data with units, use percent for readability
derappp$soil_sorption |>
  filter(substance \%in\% c("Acetamiprid", "Captan", "Copper", "Cyprodinil")) |>
  mutate(f_oc = set_units(f_oc, "\%")) |>
  select(substance, soil_pH, f_oc, Koc, Kfoc, n, sk, selected, reason) |>
  print(n = Inf)

# Show some soil degradation data with units
derappp$soil_degradation |>
  select(substance, DT50, kinetics, alpha, beta, k1, k2, g, tb, sk) |>
  print(n = 10)

# Some aquatic toxicity data with units
head(derappp$aquatic_toxicity) |>
  select(substance, formulation, derappp_species, duration, effect, level, sign, value)

# Species groupings and taxonomic IDs
derappp$species
}
\keyword{datasets}
