% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derappp.R
\docType{data}
\name{derappp_bib}
\alias{derappp_bib}
\title{List of references}
\format{
A list of \link[RefManageR:BibEntry]{RefManageR::BibEntry} objects
}
\usage{
derappp_bib
}
\description{
List of references
}
\examples{
derappp_bib[1:10]
}
\keyword{datasets}
