% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derappp.R
\docType{data}
\name{derappp_chents}
\alias{derappp_chents}
\alias{print.derappp_chents}
\title{List of chemical entities with additional information}
\format{
A list of \link[chents:chent]{chents::chent} objects, with class \code{derappp_chents}
}
\usage{
derappp_chents

\method{print}{derappp_chents}(x, n = 3, ...)
}
\arguments{
\item{x}{A list of \link[chents:chent]{chents::chent} objects, with class \code{derappp_chents}}

\item{n}{Number of entries to show}

\item{...}{For compatibility with the generic}
}
\description{
List of chemical entities with additional information
}
\examples{
# The chemical entities are stored as a list of chent objects, so we can
# access them by name
derappp_chents$Cyprodinil

derappp_chents$Myclobutanil

# The IUPAC name and source URL retrieved from the British Crop Protection
# Council (BCPC) are stored as fields in the
# `bcpc` element of the object, so we can easily access them
derappp_chents$Captan$bcpc[c("iupac_name", "source_url")]

# The PubChem information is stored as a list, so we can check which fields are available
names(derappp_chents$Captan$pubchem)

# For example, we can check the molecular formula
derappp_chents$Captan$pubchem$MolecularFormula

# We also have a print method for the complete object showing the first few
# items
print(derappp_chents, n = 2)
}
\keyword{datasets}
