% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate_substances}
\alias{translate_substances}
\alias{translate_substances.character}
\alias{translate_substances.data.frame}
\title{Translate substance names between namespaces}
\usage{
translate_substances(x, ..., from, to)

\method{translate_substances}{character}(x, substances = x, ..., from, to)

\method{translate_substances}{data.frame}(x, ..., from, to)
}
\arguments{
\item{x}{Either a character vector (see substances) or a table containing a character vector
If it is a table (a data.frame or a table derived from a data.frame such as a tibble),
there must be a column named according to the "from" argument described below.}

\item{\dots}{Currently not used}

\item{from}{Namespace of the input, defaulting to "derappp"}

\item{to}{Desired namespace of the output, defaulting to "derappp"}

\item{substances}{A character vector of substance names or primary keys
referring to one or more substance(s)}
}
\value{
A tibble with a column for each namespace, named with the namespaces
involved.
}
\description{
This function is \strong{experimental} and its interface may change in future
versions without notice.
}
\examples{
# Check which namespaces currently exist
unique(derappp$substance_keys$db)

# Translate substance primary keys from the Swiss Register of Plant Protection Products
srppp_names <- c("1-Naphthylacetic acid", "Terbuthylazine", "Pyrethrine")
srppp_pk <- c("3", "1245", "323")

# Then do the actual translation with derappp::translate_substances()
translation <- translate_substances(srppp_pk, from = "srppp")
print(translation)
translate_substances(srppp_names, from = "substance_de")

# We get warned if a substance_de is not mapped and/or translation is incomplete
translate_substances(c(srppp_names, "Glyphosate", "Kupfer"), from = "substance_de")

# There is also a method for data frames preserving all columns
input <- data.frame(x = 1:3, srppp = srppp_pk)
translate_substances(input, from = "srppp")
input_de <- data.frame(x = 1:3, substance_de = srppp_names)
translate_substances(input_de, from = "substance_de")

# We can also translate back and get multiple matches for some names
translate_substances(translation$derappp, to = "srppp")

# Or translate to another namespaces
translate_substances(srppp_pk, from = "srppp", to = "NABO_SQ")

# An example with NABO Status Quo substances
translate_substances(c("Chlorothalonil R417888", "S-Metolachlor"),
  from = "NABO_SQ")

# If we translate to a namespace that does not have a mapping for the
# substance, we get NA
translate_substances(c("Chlorothalonil R417888", "S-Metolachlor", "Diazinone"),
  from = "NABO_SQ", to = "srppp")

# We only find exact matches (room for a future extension)
translate_substances(c("S-Metolachlor", "Glyphosat", "Glyphosate"),
  from = "derappp", to = "srppp")
}
